Avoid including cryptlib.h, it's not really needed. Check if IDEA is being built or not. This is part of a large change submitted by Markus Friedl <markus@openbsd.org>

+++ ssl/ssltest.c
@@ -124,6 +124,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
+#include <inttypes.h>
 
 #define USE_SOCKETS
 #include "e_os.h"
@@ -1593,7 +1594,7 @@ static int MS_CALLBACK app_verify_callback(X509_STORE_CTX *ctx, void *arg)
 	fprintf(stderr, "In app_verify_callback, allowing cert. ");
 	fprintf(stderr, "Arg is: %s\n", (char *)arg);
 	fprintf(stderr, "Finished printing do we have a context? 0x%x a cert? 0x%x\n",
-			(unsigned int)ctx, (unsigned int)ctx->cert);
+			(uintptr_t)ctx, (uintptr_t)ctx->cert);
 	if (ctx->cert)
 		s=X509_NAME_oneline(X509_get_subject_name(ctx->cert),buf,256);
 	if (s != NULL)
